// src/modules/organizations/enums/organization-permission.enum.ts
export enum OrganizationPermission {
  ORGANIZATION_VIEW = 'ORGANIZATION_VIEW',

  ORGANIZATION_UPDATE = 'ORGANIZATION_UPDATE',

  MEMBER_VIEW = 'MEMBER_VIEW',

  MEMBER_INVITE = 'MEMBER_INVITE',

  MEMBER_UPDATE = 'MEMBER_UPDATE',

  MEMBER_REMOVE = 'MEMBER_REMOVE',

  PROCUREMENT_CREATE = 'PROCUREMENT_CREATE',

  PROCUREMENT_SUBMIT = 'PROCUREMENT_SUBMIT',

  PROCUREMENT_VIEW = 'PROCUREMENT_VIEW',

  PROCUREMENT_UPDATE = 'PROCUREMENT_UPDATE',

  PROCUREMENT_REJECT = 'PROCUREMENT_REJECT',

  PROCUREMENT_CANCEL = 'PROCUREMENT_CANCEL',

  PROCUREMENT_APPROVE = 'PROCUREMENT_APPROVE',

  PURCHASE_ORDER_CREATE = 'PURCHASE_ORDER_CREATE',

  PURCHASE_ORDER_VIEW = 'PURCHASE_ORDER_VIEW',

  PURCHASE_ORDER_APPROVE = 'PURCHASE_ORDER_APPROVE',

  FINANCE_VIEW = 'FINANCE_VIEW',

  SUPPLIER_CREATE = 'SUPPLIER_CREATE',

  SUPPLIER_VIEW = 'SUPPLIER_VIEW',

  SUPPLIER_UPDATE = 'SUPPLIER_UPDATE',

  SUPPLIER_MANAGE_CONTACTS = 'SUPPLIER_MANAGE_CONTACTS',

  SUPPLIER_MANAGE_CATEGORIES = 'SUPPLIER_MANAGE_CATEGORIES',
}